Add tests for "updating an items completed state" and "clear all comleted tasks" function
